Optimize sprite collision data storage
Prepare an array of sprite "hitmaps" that store which pixels are present instead of keeping the raw sprite graphics in memory.

This saves memory since we move from 2bpp to 1bpp, and we ignore the large chunk of required extra space in the sprite texture.

Also makes the sprite collision check simpler.